requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
tabulate>=0.9.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
redis>=5.0.0
